% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.fst_colnames_exist}
\alias{.fst_colnames_exist}
\title{Returns column names that exists in a .fst file}
\usage{
.fst_colnames_exist(path, colnames)
}
\arguments{
\item{path}{path to a .fst file}

\item{colnames}{a character vector to check}
}
\value{
column names in colnames that exists in the .fst file
}
\description{
Returns column names that exists in a .fst file
}
