% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch HILDA data}
\usage{
fetch(
  years,
  vars = NULL,
  new_varnames = NULL,
  add_population_weight = TRUE,
  add_basic_vars = TRUE,
  add_geography = FALSE,
  .dir = getOption("hildar.vault")
)
}
\value{
See \code{\link[=hil_fetch]{hil_fetch()}}.
}
\description{
Soft-deprecated, new code should use \code{\link[=hil_fetch]{hil_fetch()}}.
}
\keyword{internal}
