% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hilda_fst_dir.R
\name{get_hilda_fst_dir}
\alias{get_hilda_fst_dir}
\title{Get the directory where HILDA fst files are stored.}
\usage{
get_hilda_fst_dir()
}
\value{
a directory.
}
\description{
This function looks for \code{HILDA_FST} in global options and
the \code{.Renviron} file, in this order, and returns the first
one it finds.
}
\examples{
get_hilda_fst_dir()
}
