% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hilda_fst_path}
\alias{get_hilda_fst_path}
\title{Get a \code{HILDA_FST} environment variable}
\usage{
get_hilda_fst_path()
}
\value{
a pathname.
}
\description{
This returns an environment variable called \code{HILDA_FST} which is
the pathname where HILDA fst files are stored.
}
