% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_browse.R
\name{hil_browse}
\alias{hil_browse}
\alias{hil_crosswave_info}
\alias{hil_var_details}
\title{Browse HILDA data dictionary}
\usage{
hil_browse()

hil_crosswave_info(varname)

hil_var_details(varname, wave = 1)
}
\arguments{
\item{varname}{variable name}

\item{wave}{wave number}
}
\value{
These functions are called for their side-effect which is to
launch a relavant page on a web browser.
}
\description{
Browse HILDA data dictionary
}
