% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilda_dict}
\alias{hilda_dict}
\title{HILDA dictionary in a data.table format.}
\format{
A data table with many rows and 3 variables:
\describe{
\item{var}{(\code{character()}) variable names}
\item{wave}{(\code{list(integer())}) waves that the variable was recorded}
\item{label}{(\code{character()}) short description of the variable}
}
}
\usage{
data(hilda_dict)
}
\description{
A data.table that contains a dictionary of the loaded HILDA datasets.
}
\examples{
hilda_dict
}
\keyword{datasets}
