% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{hilda_fetch}
\alias{hilda_fetch}
\title{Fetch hilda data conditional on the args}
\usage{
hilda_fetch(years, vars = NULL, add_population_weight = TRUE,
  add_basic_vars = TRUE, add_geography = FALSE,
  responding_person_only = FALSE)
}
\arguments{
\item{vars}{a vector containing all desired variable names to be loaded.
vars can be set to "all" to fetch all columns. This may take a long time to load}

\item{waves}{wave number, multiple waves can be put into a vector}
}
\value{
a data.table object
}
\description{
Fetch hilda data conditional on the args
}
