% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_to_hildar.R
\name{hilda_to_hildar}
\alias{hilda_to_hildar}
\title{Save HILDA Stata files data to fst data}
\usage{
hilda_to_hildar(read_dir, save_dir, cores = NULL, pattern = ".dta")
}
\arguments{
\item{read_dir}{readind directory where the HILDA .dta files are}

\item{save_dir}{saving directory}

\item{cores}{number of cores}

\item{pattern}{only reads ".dta" stata files for now}
}
\description{
Save HILDA Stata files data to fst data
}
