% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hildar-package.R
\docType{package}
\name{hildar-package}
\alias{hildar}
\alias{hildar-package}
\title{hildar: HILDA Data Package and Utils Functions to Use on HILDA Survey Data}
\description{
Fast query HILDA by storing them in fst format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/hildar}
}

}
\keyword{internal}
