% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_missing_data_value}
\alias{is_missing_data_value}
\title{Is missing data value}
\usage{
is_missing_data_value(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical vector
}
\description{
Returns TRUE if the values inside a variable contains the missing data values
coding from HILDA. They usally starts with '\link{-{number}}' or a negative value
if the variable is of type numeric.
}
\examples{

h <- fetch(2011)
h[is_missing_data_value(mrcurr)]
}
