% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dict}
\alias{make_dict}
\title{Make a dictionary for a STATA dta file.}
\usage{
make_dict(stata_df)
}
\arguments{
\item{stata_df}{a data.frame read using haven::read_stata}
}
\value{
a data.table contains two columns: var and label.
}
\description{
Make a dictionary for a STATA dta file.
}
