% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_leading_numbers}
\alias{remove_leading_numbers}
\title{Remove non-missing leading numbers}
\usage{
remove_leading_numbers(x)
}
\arguments{
\item{x}{a vector or a column from HILDA data}
}
\value{
a vector
}
\description{
In HILDA all factor levels are numbered. If the leading numbers are not needed
then one can use this function to recode the factor lavels.
}
