% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_hildar}
\alias{setup_hildar}
\title{Save HILDA Stata files data to fst data}
\usage{
setup_hildar(read_dir, save_dir, n_cores = NULL, pattern = ".dta")
}
\arguments{
\item{read_dir}{readind directory where the HILDA .dta files are}

\item{save_dir}{saving directory. This will directory will be added to .Rprofile
as \code{hildar.vault}.}

\item{n_cores}{number of cores}

\item{pattern}{only reads ".dta" stata files for now}
}
\description{
Save HILDA Stata files data to fst data
}
