% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_fetch.R
\name{hil_fetch}
\alias{hil_fetch}
\title{Fetch HILDA data}
\usage{
hil_fetch(
  years = NULL,
  vars = NULL,
  new_varnames = NULL,
  add_population_weight = TRUE,
  add_basic_vars = TRUE,
  add_geography = FALSE,
  hilda_fst_dir = ifelse(!is.null(getOption("HILDA_FST")), getOption("HILDA_FST"),
    get_hilda_fst_path())
)
}
\arguments{
\item{years}{This argument allow you to specify the years of HILDA
that you like to load instead of using alphabets.
The first wave of HILDA was in 2001 and known as wave 'a',
and the following year was wave 'b'. To load multiple waves
you can be put use a numeric vector (e.g.,\code{2001:2009} would load all waves
between 2001 and 2009).}

\item{vars}{a vector containing all desired variable names to be loaded. vars can
be set to "all" to fetch all columns. This may take a long time to load}

\item{new_varnames}{a vector contains character names with its length equals length of \code{vars} to
be replaced by \code{vars} orgininal names.}

\item{add_population_weight}{take a logical value whether to add cross-sectional responding person weight
column and enumerated weight
column to the data}

\item{add_basic_vars}{take a logical value whether to add hgage (age), hgsex (sex),
mrcurr (marital status) and
hhrih (relationship in household) to the data}

\item{add_geography}{take a logical value whether to add hhsgcc (Greater statistical region)
to the data}

\item{hilda_fst_dir}{a directory where HILDA files in fst format are stored
by \code{hil_setup()}. If not given the function will check uses for 'HILDA_FST'
in your \code{.Rprofile} file first, then in \code{.Renviron} file.}
}
\value{
a data.table object
}
\description{
Fetch HILDA data from the HILDA fst files created by \code{\link[=hil_setup]{hil_setup()}}.
To set default variables for \code{\link[=hil_fetch]{hil_fetch()}} to fetch see \code{\link[=hil_user_default_vars]{hil_user_default_vars()}}.
}
\examples{
\dontrun{
summary(hil_fetch(2011))

summary(hil_fetch(2011:2012, vars = "losat"))

# Query all variables that start with 'hs' (Housing)
summary(hil_fetch(2011, vars = hil_vars("^hs")))

# Query all variables with the word 'coronavirus' in their variable description.
summary(hil_fetch(2020, vars = hil_labs("coronavirus")))
}
}
