% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{hil_fetch}
\alias{hil_fetch}
\title{Fetch HILDA data}
\usage{
hil_fetch(
  years,
  vars = NULL,
  new_varnames = NULL,
  add_population_weight = TRUE,
  add_basic_vars = TRUE,
  add_geography = FALSE,
  hilda_fst_dir = ifelse(!is.null(getOption("HILDA_FST")), getOption("HILDA_FST"),
    get_hilda_fst_path())
)
}
\arguments{
\item{years}{years (e.g. wave 1 = 2001), multiple waves can be put into a vector.}

\item{vars}{a vector containing all desired variable names to be loaded. vars can
be set to "all" to fetch all columns. This may take a long time to load}

\item{new_varnames}{a vector contains character names with its length equals length of \code{vars} to
be replaced by \code{vars} orgininal names.}

\item{add_population_weight}{take a logical value whether to add cross-sectional responding person weight
column and enumerated weight
column to the data}

\item{add_basic_vars}{take a logical value whether to add hgage (age), hgsex (sex),
mrcurr (marital status) and
hhrih (relationship in household) to the data}

\item{add_geography}{take a logical value whether to add hhsgcc (Greater statistical region)
to the data}

\item{hilda_fst_dir}{a directory where HILDA files in fst format are stored
by \code{hil_setup()}. If not given the function will check uses for 'HILDA_FST'
in your \code{.Rprofile} file first, then in \code{.Renviron} file.}
}
\value{
a data.table object
}
\description{
Fetch HILDA data
}
