% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R, R/utils.R
\name{hil_setup}
\alias{hil_setup}
\alias{make_dict}
\alias{hil_dict}
\title{Save HILDA Stata files data to fst data}
\format{
\code{hil_dict()} returns a data table with many rows and 3 variables:
\describe{
\item{var}{(\code{character()}) variable names}
\item{wave}{(\code{list(integer())}) waves that the variable was recorded}
\item{label}{(\code{character()}) short description of the variable}
}
}
\usage{
hil_setup(read_dir, save_dir)

make_dict(read_dir, save_dir = NULL)

hil_dict()
}
\arguments{
\item{read_dir}{read directory where the HILDA files that
match this \code{Combined_.*.dta} regex pattern are in.}

\item{save_dir}{a directory to save HILDA files in 'fst' format.
This directory will be added to .Rprofile as \code{hildar.vault}.}
}
\value{
\code{make_dict()} returns a data.table contains three
columns: var, label, and wave. But if \code{save_dir} is not
\code{NULL}, the dict will be saved to that location.
}
\description{
This functions looks in a directory for HILDA files in .dta.
format and save them as fst files. The fst files will be used
by \code{hil_fetch()} for loading querying HILDA data.
}
\note{
This function can take a long time to finish since each HILDA file
is quite large. One option is to use the future package to choose
your parallel backend before running \code{hil_fetch()}. The following
code chuck uses \code{multisession} which creates background R sessions
equal to the number of \code{workers}.\preformatted{library(future)
plan(multisession, workers = 2)
}
}
\examples{
# HILDA data dictionary
\dontrun{
 hil_dict()
}
}
