% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_user_default_vars.R
\name{hil_user_default_vars}
\alias{hil_user_default_vars}
\title{Return the user default HILDA variables.}
\usage{
hil_user_default_vars()
}
\value{
User default variable names as a character vector.
}
\description{
This function returns the user default HILDA variables ("HILDAR_USER_DEFAULT_VARS")
that can be found in \code{.Rprofile} and \code{.Renviron}. If 'HILDAR_USER_DEFAULT_VARS' in
both files are set, the function will give higher priority to the variables set in .Rprofile.

To edit both files, you can use \code{\link[usethis:edit]{usethis::edit_r_profile()}} and \code{\link[usethis:edit]{usethis::edit_r_environ()}}.

By default, \code{\link[=hil_fetch]{hil_fetch()}} will select the variables returned by \code{\link[=hil_user_default_vars]{hil_user_default_vars()}}
in its output.
}
\examples{
hil_user_default_vars()
}
