% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{hil_vars}
\alias{hil_vars}
\alias{hil_labs}
\title{Search HILDA data dictionary by variable or label.}
\usage{
hil_vars(pattern, ...)

hil_labs(pattern, ...)
}
\arguments{
\item{pattern}{a RegEx pattern.}
}
\value{
a character vector of variable names that match the pattern.
}
\description{
Search HILDA data dictionary by variable or label.
}
