% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_labs.R, R/hil_vars.R
\name{hil_labs}
\alias{hil_labs}
\alias{hil_vars}
\title{Search HILDA data dictionary by variable or label.}
\usage{
hil_labs(pattern, hilda_fst_dir = NULL, ...)

hil_vars(pattern, hilda_fst_dir = NULL, ...)
}
\arguments{
\item{pattern}{a RegEx pattern.}

\item{hilda_fst_dir}{a directory to HILDA fst files generated using \code{hil_setup()}.
If NULL, the result of \code{get_hilda_fst_dir()} will be used.}

\item{...}{arguments here are passed to \code{grepl()}.3}
}
\value{
a character vector of variable names that match the pattern.
}
\description{
Search HILDA data dictionary by variable or label.
}
