% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_labs.R, R/hil_vars.R
\name{hil_labs}
\alias{hil_labs}
\alias{hil_vars}
\title{Search HILDA data dictionary by variable or label.}
\usage{
hil_labs(pattern, ...)

hil_vars(pattern, ...)
}
\arguments{
\item{pattern}{a RegEx pattern.}

\item{...}{arguments here are passed to \code{grepl()}.3}
}
\value{
a character vector of variable names that match the pattern.
}
\description{
Search HILDA data dictionary by variable or label.
}
