% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hil_setup.R
\name{setup_hildar}
\alias{setup_hildar}
\title{Save HILDA Stata files data to fst data}
\usage{
setup_hildar(read_dir, save_dir, n_cores = NULL, pattern = ".dta")
}
\value{
See \code{\link[=hil_setup]{hil_setup()}}.
}
\description{
Soft-deprecated, new code should use \code{\link[=hil_setup]{hil_setup()}}.
}
\keyword{internal}
