% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardise_hilda_colnames}
\alias{standardise_hilda_colnames}
\title{Remove the leading alphabets of all the columns
except those start with 'x'}
\usage{
standardise_hilda_colnames(hilda_data)
}
\arguments{
\item{hilda_data}{a data.frame containing a HILDA dataset}
}
\value{
data.table
}
\description{
Remove the leading alphabets of all the columns
except those start with 'x'
}
